pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mdrag"
version = "0.3.0"
description = "Give any local Markdown folder a semantic-search MCP server"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Andy Lei", email = "leimingcan@icloud.com"}
]
keywords = ["mcp", "markdown", "rag", "semantic-search", "rag", "claude", "lancedb"]
classifiers = [
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Text Processing :: Markup :: Markdown",
]
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
    "lancedb>=0.14",
    "sentence-transformers>=3.0",
    "mcp>=1.0",
    "pyarrow>=16.0",
    "pathspec>=0.12",
    "watchdog>=4.0",
    "filelock>=3.12",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.6",
    "build",
    "twine",
]

[project.scripts]
mdrag = "mdrag.cli:main"

[project.urls]
Homepage = "https://github.com/andyleimc-source/mdrag"
Issues = "https://github.com/andyleimc-source/mdrag/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/mdrag"]

[tool.ruff]
line-length = 100
target-version = "py310"
